move adjustGeoCoords to utility
