+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/component/visualization/maps/openlayers/draw-styles.tsx
